requirements.txt
numpy==1.23.0
opencv_python==4.5.3.56
Pillow==9.1.1
